add: post, put delete post errors

+++ src/posts/posts.controller.ts
@@ -10,6 +10,7 @@ export class PostsController {
   @Post('/api/v1/posts')
   async create(@Body() createPostDto: CreatePostDto) {
     const createPosts = await this.postsService.create(createPostDto);
+    if(createPosts !== createPostDto) throw new Error('400: invalid input, post not created or updated')
     return createPosts;
   }
 
@@ -28,12 +29,19 @@ export class PostsController {
 
   @Put('/api/v1/posts/:id')
   async update(@Param('id') id: string, @Body() updatePostDto: UpdatePostDto) {
+    if (!this.postsService.findOne(id)){
+      
+      throw new Error('404: invalid input, post not found')
+     }
     const updatePost = await this.postsService.update(id, updatePostDto);
+    if(updatePost !== updatePostDto) throw new Error('400: invalid input, post not created or updated')
     return await updatePost;
   }
 
   @Delete('/api/v1/posts/:id')
   async remove(@Param('id') id: string) {
+    if (!this.postsService.findOne(id)) throw new Error('404: invalid input, post not found')
+    
     const deletePost = await this.postsService.remove(id);
     return await deletePost;
   }
